requirements.txt
dotenv
matplotlib
numpy
pandas
plotly
python-dotenv
requests
scikit-learn
scipy
spotipy
streamlit
tqdm
urllib3
watchdog
seleniumbase
netscape-cookies
flask
flask-ngrok